refactor: seperate type declaration parts
