+++ store/modules/main.tsx
@@ -1,53 +1,11 @@
 import { createSlice } from '@reduxjs/toolkit';
-
-export interface newRoomType {
-  address: any;
-  dealType: string;
-  deposit: number;
-  floor: string;
-  image: any;
-  latitude: number;
-  longitude: number;
-  price: number;
-  priceUnit: string;
-  recommend: number;
-  roomID: any;
-  roomSize: number;
-  roomType: string;
-  struct: string;
-}
-
-export interface newReviewType {
-  address: string;
-  regDate: number;
-  registrant: number;
-  image: any;
-  recomment: number;
-  buildingID: string;
-  landlord: number;
-  clean: number;
-  sound: number;
-  accessible: number;
-  total: number;
-  reviewID: string;
-  facility: number;
-  period: string;
-  content: string;
-}
-
-type initialStateTypes = {
-  SEARCHMODAL_OPEN: boolean;
-  SEARCHRESULT: any[];
-  SLIDEBOX: any[];
-  newRoom: newRoomType[];
-  newReview: newReviewType[];
-};
+import { MainState } from 'types';
 
 // if (typeof window !== 'undefined') {
 //     localStorage.setItem('home', 'true')
 // }
 
-const initialState: initialStateTypes = {
+const initialState: MainState = {
   SEARCHMODAL_OPEN: false,
   SEARCHRESULT: [],
   SLIDEBOX: [
