+++ types/index.ts
@@ -1,0 +1,8 @@
+export * from './filter';
+export * from './loading';
+export * from './login';
+export * from './main';
+export * from './reveiwRegister';
+export * from './room';
+export * from './roomRegister';
+export * from './tab';
