+++ types/loading.ts
@@ -1,0 +1,4 @@
+export type LoadingState = {
+  START_LOADING: boolean;
+  FINISH_LOADING: boolean;
+};
