+++ types/roomRegister.ts
@@ -1,0 +1,48 @@
+export type RoomRegisterState = {
+  COMPONENT_HANDLER: number;
+  POSTCODE_OPEN: boolean;
+  ADDRESS_OPEN: boolean;
+  longitude: null | number;
+  latitude: null | number;
+  address: string;
+  sido: string;
+  sigungu: string;
+  dong: string;
+  ri: string;
+  jibun: string;
+  detailAddress: string;
+  DETAIL_COMPONENT_OPEN: boolean;
+  roomType: string;
+  dealType: string;
+  priceUnit: string;
+  deposit: number | string;
+  price: number | string;
+  manage: 0 | 1;
+  manageCost: number | string;
+  manageSelect: Array<number>;
+  wholeFloor: number | string;
+  floor: number | string;
+  struct: string;
+  elevator: 0 | 1;
+  roomSize: number | string;
+  availFrom: string;
+  availTo: string;
+  availConsul: 0;
+  optionAll: number;
+  option: Array<number>;
+  car: 0 | 1 | 2;
+  pet: 0 | 1 | 2;
+  women: 0 | 1 | 2;
+  loan: 0 | 1 | 2;
+  title: string;
+  explain: string;
+  mainImage: string[];
+  mainImageUrl: string[];
+  room1Image: string[];
+  room1ImageUrl: string[];
+  room2Image: string[];
+  room2ImageUrl: string[];
+  postRoomSuccess: any;
+  postRoomFailure: any;
+  registrant: string;
+};
