+++ types/tab.ts
@@ -1,0 +1,8 @@
+export type TabState = {
+  TABBAR_OPEN_HOME: boolean;
+  TABBAR_OPEN_SEARCH: boolean;
+  TABBAR_OPEN_PLUS: boolean;
+  TABBAR_OPEN_CHAT: boolean;
+  TABBAR_OPEN_MYPAGE: boolean;
+  TABBAR_OPEN_POSTMODAL: boolean;
+};
